Testing: Actually cover our code, rather than everything around it.
Previously we'd see zero test coverage of `taas.py` because we were
installing it into `venv/`, which was excluded by our `.coveragerc`.

Now we're using a `--editable` install, which essentially tests the file
in-place. We've also updated our `.coveragerc` to ignore our `setup.py`
and `tests/*` directories, since we're not trying to cover them, and
including them in results is noise.