[DIAG] Better diagnostics when malformed links are encountered.
We always expect links to have a ` - ` string in them. If they didn't,
we'd get a stack-trace, but not know what field was causing problems.

With this change, we get a message back similar to:

```
ValueError: Link field ['Country / Operation'] cannot find link string ' - ' in value Ethiopia
```

This allows us to which field was doing the processing, and the data it
was using at the time.

Includes tests.