[TAAS-62] Support `--all` swith on `gss2json`
This adds support for a `--all` switch on `gss2json`, which builds all
default taxonomies. It's the same as running `gss2json` with no
switches.

The advantage of having a swith is it means that Jenkins can always run
a parameterised build, and we can just pass the parameter `--all` if we
want everything built.